requirements.txt
flask==3.0.3
flask-cors==4.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
sentence-transformers==3.0.1
google-generativeai==0.7.2
faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-dotenv==1.0.1
unstructured==0.15.12
python-docx==1.1.2
werkzeug==3.0.4
gunicorn==22.0.0
pytest==8.3.2
pytest-mock==3.14.0
pytest-flask==1.3.0
chardet<6
